Adding dockerfile in backend and frontend plus docker compose
The docker-compose.yaml is the file we're going to use when we activate using the command docker-compose. In it , docker-compose will direct to two dockerfile that is at the frontend and backend folder.  Each dockerfile is a bit similar but all it do is too start the backend and the frontend of the website. For now when docker-compose up --build is done, only the network that works, not local.